Update order.ts (#151)
amazon.com updated ui, removing total from order list view - this change switches to using the detail object's total (already present, but previously unused). It's possible that the detail's total is sometimes missing or wrong: we're gonna find out soon! If so, we should be able to either improve detail scraping OR pick the better of the main value and the detail value.