Allow reading of expected order count for a year to be updated...
...while scraping second and subsequent pages of headers.
This new behaviour might exhibit somewhat of a race condition because up
to six fetches can be concurrent, and it's possible that if a new,
increased order count is revealed in a page whose fetching and parsing has not
yet completed, the update may not increase the number of header pages to
be fetched before the fetching loop concludes.

Additionally, I yielded to the temptation to move a function from
extraction.ts to order_list_page.ts and to rename some indentifiers from
python (underscore word seperation) to camelCase.