A Breakthrough™ in sending scroll events
Achieved by finding scrollIntoView() and figuring out what level of elements
is receptive to this function (it's not the one that the debugger thinks is
wired up to scroll events.
Caution - this code still exhibits infinite looping, probably because
I've not yet wired up the carefully crafted "theWellIsDry()" in the loop
predicate.