Added a new type of test to validate topological parsing of cooked order detail html to extract payments.
In order to pass this test (which uses previously unseen html from an
amazon.com user), I've had to adjust the node classification function to
admit the possibility of 1) "ending in" prefix before last 4 card digits
instead of 4 blanked digits. 2) allowing ending in (or the blanked
digits) to be in the same element as the visible card digits.

All of the changes in payment.ts are hard to discern because I foolishly
refactored various functions at the same time (along with an amount of
function renaming). Very little existing function has been altered, but
you might struggle to see this. Sorry.