Fix-up cooked HTML handling for order_detail.payments and side-effects.
This commit was driven by order scrape tests failing due to me having
messed up how I'd started calling iframeworked.fetchURL from within
order scraping code rather than just in weird stuff like transactions.
It was seriously messing up the order tests, by expecting browser stuff
like "window" and "chrome" to be present. In order to fix this, I needed
to intercept the fetchURL flow before all of that confusion, and use the
captured html if available. Doing this interception was quite
infectious: I needed to pass request_scheduler instances all over the
place, and also pass the orderDetail URL to places that already had the
associated document (document.URL is not easily settable during
synthesis of documents, given we're doing it both in node and in
chrome).

I don't think cooked html is available in any of
the test pack's json files yet, but will be soon - I will not be surprised
if there are some test crashes when we do get some, and we'll fix them
when they arise.