Messaging no longer used to transfer available years list
...from injected (content) page to control page via background page.
Instead, the content page looks at the cached value to see if it is
present and not expired. If not, then it does the hard work to determine
the years (using an iframe, so some messaging is involved), and then
writes the result directly into cache. Validity is set to 24h.

The control page just polls the cache, updating its array of buttons
when it sees changes in the returned list from the poll, ignoring
expired entries.

Additionally, I had iframe-worker that's fetching cooked html send
keepalive messages to the background page (similar to how it does so for
transaction scraping). I did this because I suspect that some observed
failures to remove worker iframes was caused by the background page
timing out.